+++ back/prisma/data/users.ts
@@ -3,7 +3,6 @@ import { Prisma } from '@prisma/client'
 export const users: Prisma.UserCreateArgs['data'][] = [
   {
     email: 'admin@admin.com',
-    password: 'password1',
     name: 'Kevin Mamaqi',
     dni: '12345678A',
     status: 'ACTIVE',
@@ -12,7 +11,6 @@ export const users: Prisma.UserCreateArgs['data'][] = [
   },
   {
     email: 'registered@registered.com',
-    password: 'password2',
     name: 'Django Unchained',
     dni: '23456789B',
     status: 'ACTIVE',
@@ -21,7 +19,6 @@ export const users: Prisma.UserCreateArgs['data'][] = [
   },
   {
     email: 'mentor@mentor.com',
-    password: 'password3',
     name: 'Linux Mint',
     dni: '34567890C',
     status: 'ACTIVE',