+++ back/src/__tests__/resources/favorites.test.ts
@@ -1,23 +1,15 @@
 import supertest from 'supertest'
 import { expect, it, describe, beforeAll, afterAll } from 'vitest'
-import jwt, { Secret } from 'jsonwebtoken'
 import { Category } from '@prisma/client'
 import { server, testUserData } from '../globalSetup'
-import { authToken } from '../setup'
 import { prisma } from '../../prisma/client'
 import { resourceTestData } from '../mocks/resources'
 import { pathRoot } from '../../routes/routes'
 import { checkInvalidToken } from '../helpers/checkInvalidToken'
+import { authToken } from '../mocks/ssoServer'
 
 const url: string = `${pathRoot.v1.resources}/favorites`
 const categorySlug = 'testing'
-const invalidUserToken = jwt.sign(
-  { userId: 'invalid' },
-  process.env.JWT_KEY as Secret,
-  {
-    expiresIn: '1h',
-  }
-)
 
 beforeAll(async () => {
   const testCategory = (await prisma.category.findUnique({
@@ -57,24 +49,18 @@ describe('Testing GET resource/favorites/:categorySlug?', () => {
   it('Should respond 200 status with category param', async () => {
     const response = await supertest(server)
       .get(`${url}/:${categorySlug}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
     expect(response.status).toBe(200)
   })
   it('Should respond 200 status without category param', async () => {
     const response = await supertest(server)
       .get(`${url}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
     expect(response.status).toBe(200)
   })
   it('Should respond 401 status without token', async () => {
     const response = await supertest(server).get(`${url}/${categorySlug}`)
     expect(response.status).toBe(401)
-  })
-  it('Should respond 404 status without valid userId', async () => {
-    const response = await supertest(server)
-      .get(`${url}/${categorySlug}`)
-      .set('Cookie', `token=${invalidUserToken}`)
-    expect(response.status).toBe(404)
   })
 
   checkInvalidToken(`${url}/${categorySlug}`, 'get')
@@ -82,7 +68,7 @@ describe('Testing GET resource/favorites/:categorySlug?', () => {
   it('Should return favorites as an array of objects.', async () => {
     const response = await supertest(server)
       .get(`${url}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
 
     expect(response.body).toBeInstanceOf(Array)
     expect(response.body.length).toBeGreaterThan(0)
@@ -146,7 +132,7 @@ describe('Testing GET resource/favorites/:categorySlug?', () => {
     })
     const response = await supertest(server)
       .get(`${url}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
 
     expect(response.body).toBeInstanceOf(Array)
     expect(response.body.length).toBeGreaterThan(0)