+++ back/src/__tests__/seen/create.test.ts
@@ -2,11 +2,11 @@ import supertest from 'supertest'
 import { expect, describe, beforeAll, afterAll, it, afterEach } from 'vitest'
 import { Category, Resource } from '@prisma/client'
 import { server, testUserData } from '../globalSetup'
-import { authToken } from '../setup'
 import { prisma } from '../../prisma/client'
 import { pathRoot } from '../../routes/routes'
 import { resourceTestData } from '../mocks/resources'
 import { checkInvalidToken } from '../helpers/checkInvalidToken'
+import { authToken } from '../mocks/ssoServer'
 
 let testResource: Resource
 const uri = `${pathRoot.v1.seen}/`
@@ -38,7 +38,7 @@ describe('Testing viewed resource creation endpoint', () => {
   it('should mark a resource as viewed', async () => {
     const response = await supertest(server)
       .post(`${pathRoot.v1.seen}/${testResource.id}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
     const viewedResources = await prisma.viewedResource.findMany({
       where: {
         user: { dni: testUserData.admin.dni },
@@ -52,11 +52,11 @@ describe('Testing viewed resource creation endpoint', () => {
   it('should not create duplicate entries for viewed resources', async () => {
     const response = await supertest(server)
       .post(`${uri + testResource.id}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
     expect(response.statusCode).toBe(204)
     const secondResponse = await supertest(server)
       .post(`${uri + testResource.id}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
     expect(secondResponse.statusCode).toBe(204)
     const viewedResources = await prisma.viewedResource.findMany({
       where: {
@@ -70,7 +70,7 @@ describe('Testing viewed resource creation endpoint', () => {
     it('Should response status code 400 if no valid cuid is provided', async () => {
       const response = await supertest(server)
         .post(`${uri}abc`)
-        .set('Cookie', authToken.admin)
+        .set('Cookie', [`authToken=${authToken.admin}`])
       expect(response.status).toBe(400)
       expect(response.body.message[0]).toStrictEqual({
         code: 'invalid_string',
@@ -92,7 +92,7 @@ describe('Testing viewed resource creation endpoint', () => {
     it('should response status code 404 if resource does not exist', async () => {
       const response = await supertest(server)
         .post(`${uri}clnjyhw7t000008ju4ozndeqi`)
-        .set('Cookie', authToken.admin)
+        .set('Cookie', [`authToken=${authToken.admin}`])
 
       expect(response.statusCode).toBe(404)
       expect(response.body.message).toBe('Resource not found')