+++ back/src/__tests__/users/get.test.ts
@@ -2,9 +2,9 @@ import supertest from 'supertest'
 import { expect, it, describe, expectTypeOf } from 'vitest'
 import { User } from '@prisma/client'
 import { server, testUserData } from '../globalSetup'
-import { authToken } from '../setup'
 import { pathRoot } from '../../routes/routes'
 import { checkInvalidToken } from '../helpers/checkInvalidToken'
+import { authToken } from '../mocks/ssoServer'
 
 describe('Testing GET endpoint', () => {
   it('Should return error if no token is provided', async () => {
@@ -18,28 +18,28 @@ describe('Testing GET endpoint', () => {
   it('Should return successful response if valid token provided', async () => {
     const response = await supertest(server)
       .get(`${pathRoot.v1.users}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
     expect(response.status).toBe(200)
   })
 
   it('Should NOT be able to access if user level is REGISTERED', async () => {
     const response = await supertest(server)
       .get(`${pathRoot.v1.users}`)
-      .set('Cookie', authToken.user)
+      .set('Cookie', [`authToken=${authToken.user}`])
     expect(response.status).toBe(403)
   })
 
   it('Should NOT be able to access if user level is MENTOR', async () => {
     const response = await supertest(server)
       .get(`${pathRoot.v1.users}`)
-      .set('Cookie', authToken.mentor)
+      .set('Cookie', [`authToken=${authToken.mentor}`])
     expect(response.status).toBe(403)
   })
 
   it('Should be able to retrieve user info if user level is ADMIN', async () => {
     const response = await supertest(server)
       .get(`${pathRoot.v1.users}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
 
     expect(response.status).toBe(200)
     expect(response.body).toBeInstanceOf(Array)
@@ -49,7 +49,7 @@ describe('Testing GET endpoint', () => {
   it('Info retrieved by ADMIN should be of type User without password []', async () => {
     const response = await supertest(server)
       .get(`${pathRoot.v1.users}`)
-      .set('Cookie', authToken.admin)
+      .set('Cookie', [`authToken=${authToken.admin}`])
 
     type UsersWithoutPassword = Omit<User, 'password'>