+++ back/CHANGELOG.md
@@ -1,6 +1,27 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [0.4.0] - 2023-12-13
+
+### Added
+
+- `ssoUrl` to `.env` file and `appConfig`
+- implemented SSO in auth files and authenticate middleware file
+
+### Changed
+
+- Modified GET resources controller so search query param only is applied when it is 2 or more chars long.
+
+### Removed
+
+- password from User model
+
+### Testing
+
+- Added ssoServer mock
+- Updated tests to implement ssoServer mock
+- Changed token cookies to authToken
+- Changed create resource test to post
 
 ## [0.3.2] - 2023-11-23